requirements.txt
numpy==1.16.3
python-mnist==0.6
